enable live kmnet control and precise telemetry

+++ rust/crates/novasight-api/src/dto/runtime_compat.rs
@@ -238,7 +238,7 @@ pub(crate) struct StatisticsState {
     pub nvinfer_output_fps: Option<f64>,
     pub detection_batch_fps: Option<f64>,
     pub control_observation_fps: Option<f64>,
-    pub detection_data_age_ms: Option<u64>,
+    pub detection_data_age_ms: Option<f64>,
     pub telemetry_window_ms: Option<u64>,
     pub metrics_available: bool,
 }
